ecommerce/requirements.txt
pytest
pytest-django
drf-yasg

gunicorn
dj-database-url
whitenoise
django-cors-headers
